- write wrong user data registration test
