+++ tests/Fixtures/user-register.json
@@ -1,6 +1,15 @@
 {
-    "name": "u1",
-    "email": "email1@email.com",
-    "password": "123",
-    "password_confirmation": "123"
+    "1": {
+        "name": "u1",
+        "email": "email1@email.com",
+        "password": "123",
+        "password_confirmation": "123"
+    },
+
+    "2": {
+        "name": "u1",
+        "email": "email1@email.com",
+        "password": "12",
+        "password_confirmation": "123"
+    }
 }
